Separate MobiusLoopTest into individual tests
