+++ mobius-core/src/test/java/com/spotify/mobius/testdomain/Crash.java
@@ -1,0 +1,3 @@
+package com.spotify.mobius.testdomain;
+
+public class Crash implements TestEffect {}
