+++ mobius-core/src/test/java/com/spotify/mobius/testdomain/EventWithCrashingEffect.java
@@ -1,0 +1,8 @@
+package com.spotify.mobius.testdomain;
+
+public class EventWithCrashingEffect extends TestEvent {
+
+  public EventWithCrashingEffect() {
+    super("crash!");
+  }
+}
