+++ mobius-core/src/test/java/com/spotify/mobius/testdomain/EventWithSafeEffect.java
@@ -1,0 +1,8 @@
+package com.spotify.mobius.testdomain;
+
+public class EventWithSafeEffect extends TestEvent {
+
+  public EventWithSafeEffect(String id) {
+    super(id);
+  }
+}
