+++ mobius-core/src/test/java/com/spotify/mobius/testdomain/TestEffect.java
@@ -1,0 +1,3 @@
+package com.spotify.mobius.testdomain;
+
+public interface TestEffect {}
