+++ mobius-core/src/test/java/com/spotify/mobius/testdomain/TestEvent.java
@@ -1,0 +1,15 @@
+package com.spotify.mobius.testdomain;
+
+public class TestEvent {
+
+  private final String name;
+
+  public TestEvent(String name) {
+    this.name = name;
+  }
+
+  @Override
+  public String toString() {
+    return name;
+  }
+}
